fix: privatize invocation flag
