+++ google-cloud-storage/src/test/java/com/google/cloud/storage/spi/v1/AuditingHttpTransport.java
@@ -1,0 +1,46 @@
+package com.google.cloud.storage.spi.v1;
+
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.http.LowLevelHttpRequest;
+import com.google.api.client.http.LowLevelHttpResponse;
+import com.google.cloud.Tuple;
+import com.google.common.collect.ImmutableList;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public final class AuditingHttpTransport extends HttpTransport {
+  private final LowLevelHttpResponse response;
+  private final List<Tuple<String, String>> buildRequestCalls;
+  private final List<Tuple<String, String>> addHeaderCalls;
+
+  public AuditingHttpTransport(LowLevelHttpResponse response) {
+    this.response = response;
+    this.buildRequestCalls = Collections.synchronizedList(new ArrayList<>());
+    this.addHeaderCalls = Collections.synchronizedList(new ArrayList<>());
+  }
+
+  public List<Tuple<String, String>> getBuildRequestCalls() {
+    return ImmutableList.copyOf(buildRequestCalls);
+  }
+
+  public List<Tuple<String, String>> getAddHeaderCalls() {
+    return ImmutableList.copyOf(addHeaderCalls);
+  }
+
+  @Override
+  protected LowLevelHttpRequest buildRequest(String method, String url) {
+    buildRequestCalls.add(Tuple.of(method, url));
+    return new LowLevelHttpRequest() {
+      @Override
+      public void addHeader(String name, String value) {
+        addHeaderCalls.add(Tuple.of(name, value));
+      }
+
+      @Override
+      public LowLevelHttpResponse execute() {
+        return response;
+      }
+    };
+  }
+}
